airflow/requirements.txt
apache-airflow-providers-mysql
apache-airflow-providers-docker
PyMySQL
pandas
kagglehub